move removing field from Info.plist to EasDispatch
