+++ apps/eas-expo-go/scripts/eas-build-pre-install.sh
@@ -39,5 +39,5 @@ pushd $ROOT_DIR/tools
 yarn
 
 if [ "$EAS_BUILD_PROFILE" = "release-client" ] && [ "$EAS_BUILD_PLATFORM" = "ios" ]; then
-  et remove-background-permissions-from-info-plists
+  et eas remove-background-permissions-from-info-plist
 fi
